# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shri BalaJi Dham HMS API",
        "description": "Hotel Management System REST API for Shri BalaJi Dham Hotel, Mathura. Provides endpoints for room management, bookings, guest management, financials, and more.",
        "contact": {
            "name": "Shri BalaJi Dham Support",
            "email": "sribalajidhamhotel@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "List all properties",
                "operationId": "listProperties",
                "responses": {
                    "200": {
                        "description": "Properties list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "SBD-MTR"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Shri BalaJi Dham Hotel"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Mathura"
                                                    },
                                                    "country": {
                                                        "type": "string",
                                                        "example": "India"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/properties/{id}": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get property details",
                "operationId": "showProperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Paste the token from /auth/login response (without \"Bearer\" prefix)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}